Commit 04/08

+++ Models/ServicoCobranca.cs
@@ -4,11 +4,11 @@ namespace Conta_Certa.Models;
 public class ServicoCobranca
 {
     public long IdServicoCobranca { get; }
-    public int IdCobranca { get; }
+    public long IdCobranca { get; }
     public Servico Servico { get; }
     public int Quantidade { get; set; } = 0;
 
-    public ServicoCobranca(long idServicoCobranca, int idCobranca, Servico servico, int quantidade)
+    public ServicoCobranca(long idServicoCobranca, long idCobranca, Servico servico, int quantidade)
     {
         IdServicoCobranca = idServicoCobranca;
         IdCobranca = idCobranca;
